add game/player state repository
