+++ server/src/app.module.ts
@@ -3,6 +3,7 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { ServicesModule } from './services/services.module';
 import { GatewaysModule } from './gateways/gateways.module';
+import { RepositoriesModule } from './repositories/repositories.module';
 import SampleConfig from './configs/sample.config';
 
 @Module({
@@ -11,6 +12,7 @@ import SampleConfig from './configs/sample.config';
     mongooseConfig,
     ServicesModule,
     GatewaysModule,
+    RepositoriesModule,
   ],
 })
 export class AppModule {}
