+++ server/src/gateways/gateways.module.ts
@@ -1,3 +1,4 @@
+import { RepositoriesModule } from './../repositories/repositories.module';
 import { Module } from '@nestjs/common';
 import { ServicesModule } from 'src/services/services.module';
 import { ChattingGateway } from './chatting.gateway';
@@ -7,7 +8,7 @@ import { ConfigModule } from '@nestjs/config';
 import { GameGateway } from './game.gateway';
 
 @Module({
-  imports: [ServicesModule, ConfigModule],
+  imports: [ServicesModule, RepositoriesModule, ConfigModule],
   providers: [ChattingGateway, TradeGateway, JoinGateway, GameGateway],
 })
 export class GatewaysModule {}
