+++ server/src/gateways/item.gateway.ts
@@ -1,4 +1,3 @@
-import { GameService } from 'src/services/game.service';
 import {
   SubscribeMessage,
   WebSocketGateway,
@@ -6,21 +5,16 @@ import {
 } from '@nestjs/websockets';
 import { Server } from 'socket.io';
 import { ItemRequestDto } from 'src/dto/item-request.dto';
-import { EffectResponse, EffectService } from 'src/services/effect.service';
 import { ItemService } from 'src/services/item.service';
 
-import { ITEM_REQUEST, ITEM_RESPONSE } from './events';
+import { ITEM_REQUEST } from './events';
 
 @WebSocketGateway()
 export class GameGateway {
   @WebSocketServer()
   private server: Server;
 
-  constructor(
-    private itemService: ItemService,
-    private gameService: GameService,
-    private effectService: EffectService,
-  ) {}
+  constructor(private itemService: ItemService) {}
 
   @SubscribeMessage(ITEM_REQUEST)
   async receiveItemRequest(client: any, payload: ItemRequestDto) {
