+++ server/src/gateways/join.gateway.ts
@@ -21,9 +21,9 @@ import {
 import { Server, Socket } from 'socket.io';
 import { Player, PlayerInfo, PlayerStatus } from 'src/schemas/player.schema';
 import { PlayerService } from 'src/services/player.service';
-import { GameService } from 'src/services/game.service';
 import { Game } from 'src/schemas/game.schema';
 import { Types } from 'mongoose';
+import { GameRepository } from 'src/repositories/game.repository';
 
 @WebSocketGateway()
 export class JoinGateway implements OnGatewayConnection, OnGatewayDisconnect {
@@ -33,9 +33,9 @@ export class JoinGateway implements OnGatewayConnection, OnGatewayDisconnect {
   private logger: Logger = new Logger('AppGateway');
 
   constructor(
+    private gameRepository: GameRepository,
     private playerService: PlayerService,
     private joinService: JoinService,
-    private gameService: GameService,
   ) {}
 
   handleConnection(client: Socket): void {
@@ -92,7 +92,7 @@ export class JoinGateway implements OnGatewayConnection, OnGatewayDisconnect {
         }
 
         const nowDate: Date = new Date();
-        const nextDate: Date = this.gameService.getNextDate(game._id);
+        const nextDate: Date = this.gameRepository.getNextDate(game._id);
         const dateDiff: number = nextDate.getTime() - nowDate.getTime();
 
         this.server
@@ -113,7 +113,7 @@ export class JoinGateway implements OnGatewayConnection, OnGatewayDisconnect {
       _id: playerId,
       name,
       room,
-    } = await this.playerService.createDocument({
+    } = await this.playerService.create({
       ...payload,
       clientId: client.id,
       status: 'connected',
@@ -211,7 +211,7 @@ export class JoinGateway implements OnGatewayConnection, OnGatewayDisconnect {
         statuses: this.getStatuses('play'),
       });
 
-      this.gameService.createGame(gameInfo.gameId, room);
+      this.gameRepository.createGameState(gameInfo.gameId, room);
     } else {
       this.server.to(room).emit(CHATTING_SERVER_MESSAGE, {
         user: '관리자',
