+++ server/src/repositories/repositories.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { GameRepository } from './game.repository';
+import { PlayerRepository } from './player.repository';
+
+@Module({
+  providers: [GameRepository, PlayerRepository],
+  exports: [GameRepository, PlayerRepository],
+})
+export class RepositoriesModule {}
