+++ server/src/services/trade.service.ts
@@ -4,10 +4,10 @@ import { Model, Types } from 'mongoose';
 import { TradeCancelDto } from 'src/dto/trade-cancel.dto';
 import { TradeRequestDto } from 'src/dto/trade-request.dto';
 import { TradeResponseDto } from 'src/dto/trade-response.dto';
+import { GameRepository, TimeState } from 'src/repositories/game.repository';
 import { Player, PlayerDocument } from 'src/schemas/player.schema';
 import { Stock, StockDocument } from 'src/schemas/stock.schema';
 import { Trade, TradeDocument } from 'src/schemas/trade.schema';
-import { GameService, TimeState } from './game.service';
 
 @Injectable()
 export class TradeService {
@@ -15,7 +15,7 @@ export class TradeService {
     @InjectModel(Trade.name) private tradeModel: Model<TradeDocument>,
     @InjectModel(Stock.name) private stockModel: Model<StockDocument>,
     @InjectModel(Player.name) private playerModel: Model<PlayerDocument>,
-    private gameService: GameService,
+    private gameRepository: GameRepository,
   ) {}
 
   async handleTrade(
@@ -24,7 +24,7 @@ export class TradeService {
     const { playerId, gameId, week, day, tick, corpId, price, quantity, deal } =
       tradeRequestDto;
 
-    const time: TimeState = this.gameService.getTime(gameId);
+    const time: TimeState = this.gameRepository.getTime(gameId);
 
     if (
       time.week !== week ||
@@ -259,23 +259,6 @@ export class TradeService {
     tradeCancelDto: TradeCancelDto,
   ): Promise<TradeResponseDto> {
     const { playerId, corpId, _id } = tradeCancelDto;
-
-    // const time: TimeState = this.gameService.getTime(gameId);
-
-    // if (
-    //   time.week !== week ||
-    //   time.day !== day ||
-    //   time.tick !== tick ||
-    //   time.tick < 1 ||
-    //   time.tick > 3 ||
-    //   day < 1
-    // ) {
-    //   const timeError = new Error(
-    //     '거래 시간이 불일치하거나 거래 불가능 시간입니다.',
-    //   );
-    //   timeError.name = 'TimeException';
-    //   throw timeError;
-    // }
 
     const player = await this.playerModel.findOne({
       _id: Types.ObjectId(playerId),
